Really the last step: Add component for ingress DNS record
